Replace empty bookmark title with best guess
Some users remove the title and only keep the favicon when adding
links to the bookmark toolbar, to improve appearance and save space.

The bookmark library and sidebar will come up with a best guess
for a bookmark title, if it's left empty. The title is constructed
using the URI.

Added this default Firefox behavior to Bookmark Search Plus 2.