Fixed many errors
I came to fix a few errors.  Found many more than expected, and I'm not one to do a job half-way! :)

There are a few words that were unrecognizable, so I didn't hazard a guess on those! ;)

I don't have privileges to update the wiki, so you may want to update that.